Add lane information on match row, fix text cutoff
Add text showing which lane a hero was in, or indicate they were roaming,
to match rows in tables on Overview and Matches player pages.
Additionally, the ">" next to the hero name was getting cut off by
the parent container when the hero was Legion Commander, so widen
the container slightly.