Added lane outcome to the hero tooltips!
Fixed #1655 per user request. Waving the mouse over a hero's icon in the match overview's `BuildingMap` component will now have a new line at the bottom of the tooltip pop-up to say either "Win", "Loss", or "No result" based on the outcome of that lane. It is calculated the same way as in the LaneStory class contained in the MatchStory.jsx file.

It is probably better practice to take that part of the LaneStory class out into its own file so that we can refactor and stick to the DRY principle and use indirection, but I don't know what your rules are about adding new files or components, so I just played it safe and ported the functionality directly into this component via hard coding.

I hope this PR request is worthy of acceptance!