Convert and consume existing translations to new format
Convert existing translations to use {num} for template instead of %s/d.  Update all uses of util.format to utilize formatTemplateToString.  formatTemplatetoString is a function that automatically converts the return of formatTemplate to a string (rather than the array that’s normally returned).  This ensures closer compatibility with util.format